♻️ refactor: #24 #27 Avatar画像の処理ロジックを抽出
